refactor: extract common component in its own file
