+++ src/features/search/pages/SuggestedPlacesOrVenues/HiddenAccessibleResultNumber.tsx
@@ -1,0 +1,24 @@
+import React from 'react'
+
+import { AccessibilityRole } from 'libs/accessibilityRole/accessibilityRole'
+import { plural } from 'libs/plural'
+import { HiddenAccessibleText } from 'ui/components/HiddenAccessibleText'
+
+export const HiddenAccessibleResultNumber = ({
+  nbResults,
+  show,
+}: {
+  nbResults: number
+  show: boolean
+}) => {
+  const numberOfResults = plural(nbResults, {
+    one: '# résultat',
+    other: '# résultats',
+  })
+
+  return (
+    <HiddenAccessibleText accessibilityRole={AccessibilityRole.STATUS}>
+      {show ? numberOfResults : ''}
+    </HiddenAccessibleText>
+  )
+}
